samples/summarizer/requirements.txt
streamlit>=1.32.0
webvtt-py>=0.4.6
openai>=1.12.0
tiktoken>=0.6.0 